Updated Compare Widget:
Updated `_refresh_table` to do a "partial refresh" of sorts when there are no matches
    We loop over the entire datatable, and for each cell, we update the rich.Text.
    Specifically, we remove space characters `" "` from the beginning and end of each line, and this removes the background color from the cells.
    We need do this because its possible some cells were previously matches but are now unmatched, and so, they should not have the background color.

Also did some code cleanup/refactor in the "full refresh" part of the `_refresh_table`

Added `_get_merkle_from_row_key` which basically takes a row_key as input and returns the corresponding merkle from self.submerkle.children. This is basically a helper/util function.

Added `_get_row_key_from_scroll_y` which is pretty cool. Given a `scroll_y` value, it will return the row key of the first visible row in the data table. To achieve this, we use the the internal attribute `_y_offsets` of DataTable.

Updated `watch_scroll_y`, so that we only sync scroll when there are matches on-screen.
Once we scroll past matches, and into unmatched merkles, we dont want to sync scroll anymore.