Style changes to make CompareWidgets flush against each other, feels nicer
We'd like FilePicker to have a nice horizontal margin of 2, which makes them flush with the Input at the top on the left and right edges of the screen, but gives a nice spacing of 2 between the two FilePickers.

But the same doesn't feel nice for CompareWidget. We'd still want the margin of 2 on the left and right edges of screen, to make the CompareWidgets flush against the Input at the top, but, we dont want spacing between the two CompareWidgets.

Also, adding a margin of 2 on the right for the Horizontal label, so that the label doesnt show on top of the scrollbar portion of CompareWidget, and is just displayed above the actual DataTable columns.